Revert "Don't update file records before bumping BN"
This reverts commit 92f1216f53311519092c136f1323bf67a4b06603.

The commit fixed the problem that listeners of "eleventy.before" event
might call `updateFileRecord()` before bumping the build number.

However, the fix also introduced a problem that a listener of
"eleventy.after" event might call `updateFileRecord()` and it may be
handled as file update in the next build instead of the current build.