Invoke incrementBuildNumber() first of all
incrementBuildNumber() is called on "eleventy.before" event.

Listeners of "eleventy.before" event will be called in the order they
were added.

However, incrementBuildNumber() should be called before any of the
other "eleventy.before" listeners, otherwise calling
`updateFileRecord()` in those listeners will update file records with
the previous build number and won't work correctly.

To make incrementBuildNumber() be called before any other listeners, by
this commit, the plugin calls `eleventyConfig.events.prependListener()`
instead of calling `eleventyConfig.on()` to add the listener which
invokes incrementBuildNumber(), and make the listener sync function.

This is the replacement for the incomplete fix in v1.1.1.